[Feature #156461691] Modularise models in different packages
