+++ api/equipment/schema.py
@@ -4,7 +4,7 @@ from graphene import relay, Schema
 from graphene_sqlalchemy import (SQLAlchemyObjectType, 
                                  SQLAlchemyConnectionField)
 
-from equipment.models import Equipment as EquipmentModel
+from api.equipment.models import Equipment as EquipmentModel
 
 
 class Equipment(SQLAlchemyObjectType):
