+++ api/floor/models.py
@@ -5,11 +5,12 @@ from sqlalchemy.orm import relationship
 
 from helpers.database import Base
 from utilities.utility import Utility
-from room.models import Room
+from api.block.models import Block
 
 
-class Equipment(Base, Utility):
-    __tablename__ = 'equipments'
+class Floor(Base, Utility):
+    __tablename__ = 'floors'
     id = Column(Integer, primary_key=True)
     name = Column(String, nullable=False)
-    room_id = Column(Integer, ForeignKey('rooms.id'))
+    block_id = Column(Integer, ForeignKey('blocks.id'))
+    room = relationship('Room')
