+++ api/room/models.py
@@ -5,12 +5,14 @@ from sqlalchemy.orm import relationship
 
 from helpers.database import Base
 from utilities.utility import Utility
-from location.models import Location
+from api.floor.models import Floor
 
 
-class Block(Base, Utility):
-    __tablename__ = 'blocks'
+class Room(Base, Utility):
+    __tablename__ = 'rooms'
     id = Column(Integer, primary_key=True)
     name = Column(String, nullable=False)
-    location_id = Column(Integer, ForeignKey('locations.id'))
-    floor = relationship('Floor')
+    room_type = Column(String, nullable=False)
+    capacity = Column(Integer, nullable=False)
+    floor_id = Column(Integer, ForeignKey('floors.id'))
+    equipment = relationship('Equipment')
