+++ helpers/database.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm import scoped_session, sessionmaker
 
 from config import config
 
-config_name = os.getenv('FLASK_CONFIG')
+config_name = os.getenv('APP_SETTINGS')
 database_uri = config.get(config_name).SQLALCHEMY_DATABASE_URI
 engine = create_engine(database_uri, convert_unicode=True)
 db_session = scoped_session(sessionmaker(autocommit=False,
