+++ manage.py
@@ -5,7 +5,7 @@ from flask_script import Manager, Shell
 from app import create_app
 
 
-app = create_app(os.getenv('FLASK_CONFIG') or 'default')
+app = create_app(os.getenv('APP_SETTINGS') or 'default')
 manager = Manager(app)
 
 def make_shell_context():
